Working examples for HOAS
There is a bit of wrapping/unwrapping literals required given a function
is

Value => Expression

and apply is

(Function, Expression)

So if you have nested applications like

fun { a =>
  foo(a)
}

we need to wrap a in a Literal to turn it back into an Expression. Still
undecided if this is a bad thing or not.